requirements.txt
streamlit
pandas
mysql-connector-python
fastapi
uvicorn
pydantic
requests